requirements.txt
RPi.GPIO==0.7.0
smbus2==0.4.1
simple_pid==1.0.1
redis==3.5.3
pymongo==3.4.0
websockets==9.0.1
flask==1.1.2
